pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "majordomus"
version = "0.1.0"
description = "Major Domus v1 - Multi-Project Governance Orchestrator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5",
  "mypy>=1.8",
  "types-PyYAML",
]

[project.scripts]
majordomus = "majordomus.cli.main:entrypoint"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
majordomus = ["schemas/*.json"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = [
  "E203",  # совместимость со стилями форматтера
]

[tool.ruff.lint.isort]
known-first-party = ["majordomus"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
